Moved TokenRequest* classes from the sample app to azure-identity.
