+++ sdk/identity/azure-identity/build.gradle
@@ -22,4 +22,8 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
+    implementation project(":sdk:core:azure-core")
+    implementation "androidx.lifecycle:lifecycle-extensions:$lifecycleVersion"
+    implementation "com.jakewharton.threetenabp:threetenabp:$threeTenAbpVersion"
+    implementation "com.squareup.okhttp3:okhttp:$okHttpVersion"
 }
